component organization (#54)
