+++ src/components/README.md
@@ -1,0 +1,13 @@
+# Components
+
+## Common
+
+Generalized components.
+
+## Pages
+
+Pages that compose the common components.
+
+## Shell
+
+Shell components "outside" of the page itself (router, navbar, sidebar, background, etc).
